+++ head/lib/Head/Controller/UiDevices.pm
@@ -126,10 +126,9 @@ sub devicepost {
 
       $results->finish;
 
-      ## email_notify field in devices is deprecated
       $db->query("INSERT INTO devices \
-(name, devices.desc, create_time, ip, mac, no_dhcp, rt, defjump, speed_in, speed_out, qs, limit_in, sum_limit_in, profile, notified, blocked, bot, email_notify, client_id) \
-VALUES (?, ?, NOW(), ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 0, 0, 1, 0, ?)",
+(name, devices.desc, create_time, ip, mac, no_dhcp, rt, defjump, speed_in, speed_out, qs, limit_in, sum_limit_in, profile, notified, blocked, bot, client_id) \
+VALUES (?, ?, NOW(), ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, 0, 0, 0, ?)",
         $j->{name},
         $j->{desc} // '',
         scalar($ipo->numeric),